admin-panel/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9